project is ready and also one game, needs the second application
