+++ Alien_Attack/Game/img/ally.gif
