+++ Alien_Attack/Game/img/laserAlly.gif
