+++ README.md
@@ -1,1 +1,21 @@
 # RicardGame
+
+Welcome to a implementation of Software Product Line for Professor Arturo final project
+
+We implemented an SPL for a space invader's kind of game using pygame library. Our goal is to apply the theory of 
+Object Oriented SPL in games. Using pygame we created interfaces to be instantiated by the developer which will 
+focus on a 2D game shooter.
+
+The interfaces will be found inside the LPS folder for reference to create the components.
+
+We also have two implementations in both Alien_descend and _____________ folders in which our components were created
+on the Components.py file.
+
+All of our used sprites, images and sounds are stored inside the Game folder, along with some images for the developer
+to use.
+
+Thank you,
+
+Ricardo Fragoso(@Karagusto)
+
+
