+++ Tutorial/scene6.py
